+++ assets/js/script.js
@@ -1,3 +1,7 @@
+var Shape = require('./shape.js')
+var Pikachu = require('./shape.js')
+var ball = require('./ball')
+
 $(function () {
   // HELPER VARS AND FUNCTIONS
   var canvas = document.getElementById('canvas')
@@ -10,53 +14,12 @@ $(function () {
 
   // CONSTRUCTOR & PROTOTYPE
   // Main Shape Contructor & Proto
-  function Shape (positionX, positionY, color, width = 20, height = 20) {
-    this.positionX = positionX
-    this.positionY = positionY
-    this.color = color
-    this.width = width
-    this.height = height
-    this.edgeLeft = 0
-    this.edgeRight = canvasWidth - this.width
-    this.edgeMiddle = (canvasWidth - gutter) / 2
-    this.edgeBottom = canvasHeight - this.height
-    this.edgeTop = 0 + this.height
-  }
-  Shape.prototype.moveTo = function (x, y) {
-    this.positionX += x
-    this.positionY += y
-    console.log(y, this.positionY)
-  }
+
 
   // Main Shape Pikachu & Proto
-  function Pikachu (positionX, positionY, color) {
-    Shape.call(this, positionX, positionY, color, 100, 100)
-    this.speed = 200
-  }
 
-  // TODO: Find a better solution to check middle collision
-  Pikachu.prototype = Object.create(Shape.prototype)
-  Pikachu.prototype.constructor = Shape
-
-  Pikachu.prototype.moveLeft = function (modifier) {
-    if (this.positionX <= this.edgeLeft) return
-    this.moveTo(-(this.speed * modifier), 0)
-  }
-  Pikachu.prototype.moveRight = function (modifier) {
-    if (this.positionX >= this.edgeRight) return
-    this.moveTo(this.speed * modifier, 0)
-  }
 
   // Main Shape Ball & Proto
-  function Ball (positionX, positionY, color) {
-    Shape.call(this, positionX, positionY, color)
-    this.radius = 20
-    this.startAngle = 0
-    this.endAngle = 360
-    this.speed = 100
-  }
-  Ball.prototype = Object.create(Shape.prototype)
-  Ball.prototype.constructor = Shape
 
   // Ball.prototype.checkColission = function () {
   //   if (this.positionY >= this.edgeBottom) this.hitBottom = true
